feat: Projeto completo com todas as aplicações integradas
🎯 Sistema Completo de Condomínio:
- admin-laravel: Painel administrativo completo
- app-vigilante: Aplicação mobile para vigilantes
- app-morador: Aplicação PWA para moradores
- relatorio-camera: Sistema de relatórios

🔗 Integrações Realizadas:
- Autenticação centralizada entre todas as apps
- Sistema de escalas sincronizado
- API endpoints funcionais
- Dados compartilhados via MySQL

📱 App Vigilante:
- Dashboard com escalas efetivas
- Integração com EscalaDiaria
- Indicadores visuais de ajustes
- Navegação entre postos

🏠 App Morador:
- Sistema PWA funcional
- Autenticação com tabela usuario
- Interface responsiva

📊 Seeders e Dados:
- Dados completos de exemplo
- Estrutura de testes funcionais
- Credenciais documentadas

📚 Documentação:
- README completo
- Arquivos de documentação técnica
- Guias de correções aplicadas

+++ app-morador/database/migrations/2025_08_28_164716_create_solicitacao_panicos_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('localizacao')->nullable();
             $table->json('coordenadas')->nullable();
             $table->timestamp('atendido_em')->nullable();
-            $table->unsignedBigInteger('atendido_por')->nullable(); // ID do vigilante
+            $table->unsignedInteger('atendido_por')->nullable(); // ID do vigilante
             $table->text('observacoes_atendimento')->nullable();
             $table->timestamps();
